Backend config update and new frontend data-models pages
